pulling maps out of engine_validation.
